Get Citoid data

Get Citoid data

cit_oid(id, format = "mediawiki", accept_language = NULL, ...)

cit_oid_(id, format = "mediawiki", accept_language = NULL, ...)

Arguments

id
(character) id of an article, DOI, ISBN, PMCID or PMID

format
(character) the format to use for the resulting citation data. one of mediawiki (default), zotero, mediawiki-basefields, bibtex

accept_language
(character) for some articles the result depends on the accept language value, so provide it if localized content is required

...
curl options passed on to crul::verb-GET

Value

list of lists or character, see http://opencitations.net/index/coci/api/v1 for explanation of the resulting columns

Details

cit_oid_() gets raw text (either bibtex or JSON), and cit_oid() parses the text as appropriate for the type

References

https://en.wikipedia.org/api/rest_v1/#!/Citation/getCitation, https://www.mediawiki.org/wiki/Citoid

Examples

url<-"https://en.wikipedia.org/api/rest_v1/data/citation/mediawiki/30446726"
if (crul::ok(url)) {
  pmid1 <- 30446726
  cit_oid(pmid1)
}

#> [[1]]
#> [[1]]$key
#> [1] "7MG9TUAZ"
#> 
#> [[1]]$version
#> [1] 0
#> 
#> [[1]]$itemType
#> [1] "journalArticle"
#> 
#> [[1]]$tags
#> [[1]]$tags[[1]]
#> [[1]]$tags[[1]]$tag
#> [1] "Animals"
#> 
#> [[1]]$tags[[1]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[2]]
#> [[1]]$tags[[2]]$tag
#> [1] "Antigen Presentation"
#> 
#> [[1]]$tags[[2]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[3]]
#> [[1]]$tags[[3]]$tag
#> [1] "Antigens"
#> 
#> [[1]]$tags[[3]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[4]]
#> [[1]]$tags[[4]]$tag
#> [1] "CD4-Positive T-Lymphocytes"
#> 
#> [[1]]$tags[[4]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[5]]
#> [[1]]$tags[[5]]$tag
#> [1] "Cell Proliferation"
#> 
#> [[1]]$tags[[5]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[6]]
#> [[1]]$tags[[6]]$tag
#> [1] "Cells, Cultured"
#> 
#> [[1]]$tags[[6]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[7]]
#> [[1]]$tags[[7]]$tag
#> [1] "Coculture Techniques"
#> 
#> [[1]]$tags[[7]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[8]]
#> [[1]]$tags[[8]]$tag
#> [1] "Female"
#> 
#> [[1]]$tags[[8]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[9]]
#> [[1]]$tags[[9]]$tag
#> [1] "Humans"
#> 
#> [[1]]$tags[[9]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[10]]
#> [[1]]$tags[[10]]$tag
#> [1] "Lectins, C-Type"
#> 
#> [[1]]$tags[[10]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[11]]
#> [[1]]$tags[[11]]$tag
#> [1] "Lung"
#> 
#> [[1]]$tags[[11]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[12]]
#> [[1]]$tags[[12]]$tag
#> [1] "Mice"
#> 
#> [[1]]$tags[[12]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[13]]
#> [[1]]$tags[[13]]$tag
#> [1] "Mice, Inbred BALB C"
#> 
#> [[1]]$tags[[13]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[14]]
#> [[1]]$tags[[14]]$tag
#> [1] "Mucous Membrane"
#> 
#> [[1]]$tags[[14]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[15]]
#> [[1]]$tags[[15]]$tag
#> [1] "Mycobacterium bovis"
#> 
#> [[1]]$tags[[15]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[16]]
#> [[1]]$tags[[16]]$tag
#> [1] "Mycobacterium tuberculosis"
#> 
#> [[1]]$tags[[16]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[17]]
#> [[1]]$tags[[17]]$tag
#> [1] "Programmed Cell Death 1 Receptor"
#> 
#> [[1]]$tags[[17]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[18]]
#> [[1]]$tags[[18]]$tag
#> [1] "Receptors, Immunologic"
#> 
#> [[1]]$tags[[18]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[19]]
#> [[1]]$tags[[19]]$tag
#> [1] "Tuberculosis"
#> 
#> [[1]]$tags[[19]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[20]]
#> [[1]]$tags[[20]]$tag
#> [1] "Tuberculosis Vaccines"
#> 
#> [[1]]$tags[[20]]$type
#> [1] 1
#> 
#> 
#> [[1]]$tags[[21]]
#> [[1]]$tags[[21]]$tag
#> [1] "Vaccination"
#> 
#> [[1]]$tags[[21]]$type
#> [1] 1
#> 
#> 
#> 
#> [[1]]$title
#> [1] "Enhanced protection conferred by mucosal BCG vaccination associates with presence of antigen-specific lung tissue-resident PD-1+ KLRG1- CD4+ T cells"
#> 
#> [[1]]$pages
#> [1] "555–564"
#> 
#> [[1]]$ISSN
#> [[1]]$ISSN[[1]]
#> [1] "1935-3456"
#> 
#> 
#> [[1]]$journalAbbreviation
#> [1] "Mucosal Immunol"
#> 
#> [[1]]$publicationTitle
#> [1] "Mucosal Immunology"
#> 
#> [[1]]$volume
#> [1] "12"
#> 
#> [[1]]$issue
#> [1] "2"
#> 
#> [[1]]$date
#> [1] "2019-03"
#> 
#> [[1]]$abstractNote
#> [1] "BCG, the only vaccine licensed against tuberculosis, demonstrates variable efficacy in humans. Recent preclinical studies highlight the potential for mucosal BCG vaccination to improve protection. Lung tissue-resident memory T cells reside within the parenchyma, potentially playing an important role in protective immunity to tuberculosis. We hypothesised that mucosal BCG vaccination may enhance generation of lung tissue-resident T cells, affording improved protection against Mycobacterium tuberculosis. In a mouse model, mucosal intranasal (IN) BCG vaccination conferred superior protection in the lungs compared to the systemic intradermal (ID) route. Intravascular staining allowed discrimination of lung tissue-resident CD4+ T cells from those in the lung vasculature, revealing that mucosal vaccination resulted in an increased frequency of antigen-specific tissue-resident CD4+ T cells compared to systemic vaccination. Tissue-resident CD4+ T cells induced by mucosal BCG displayed enhanced proliferative capacity compared to lung vascular and splenic CD4+ T cells. Only mucosal BCG induced antigen-specific tissue-resident T cells expressing a PD-1+ KLRG1- cell-surface phenotype. These cells constitute a BCG-induced population which may be responsible for the enhanced protection observed with IN vaccination. We demonstrate that mucosal BCG vaccination significantly improves protection over systemic BCG and this correlates with a novel population of BCG-induced lung tissue-resident CD4+ T cells."
#> 
#> [[1]]$DOI
#> [1] "10.1038/s41385-018-0109-1"
#> 
#> [[1]]$extra
#> [1] "PMID: 30446726\nPMCID: PMC7051908"
#> 
#> [[1]]$libraryCatalog
#> [1] "PubMed"
#> 
#> [[1]]$url
#> [1] "https://pubmed.ncbi.nlm.nih.gov/30446726"
#> 
#> [[1]]$accessDate
#> [1] "2021-07-17"
#> 
#> [[1]]$author
#> [[1]]$author[[1]]
#> [[1]]$author[[1]][[1]]
#> [1] "N. C."
#> 
#> [[1]]$author[[1]][[2]]
#> [1] "Bull"
#> 
#> 
#> [[1]]$author[[2]]
#> [[1]]$author[[2]][[1]]
#> [1] "E."
#> 
#> [[1]]$author[[2]][[2]]
#> [1] "Stylianou"
#> 
#> 
#> [[1]]$author[[3]]
#> [[1]]$author[[3]][[1]]
#> [1] "D. A."
#> 
#> [[1]]$author[[3]][[2]]
#> [1] "Kaveh"
#> 
#> 
#> [[1]]$author[[4]]
#> [[1]]$author[[4]][[1]]
#> [1] "N."
#> 
#> [[1]]$author[[4]][[2]]
#> [1] "Pinpathomrat"
#> 
#> 
#> [[1]]$author[[5]]
#> [[1]]$author[[5]][[1]]
#> [1] "J."
#> 
#> [[1]]$author[[5]][[2]]
#> [1] "Pasricha"
#> 
#> 
#> [[1]]$author[[6]]
#> [[1]]$author[[6]][[1]]
#> [1] "R."
#> 
#> [[1]]$author[[6]][[2]]
#> [1] "Harrington-Kandt"
#> 
#> 
#> [[1]]$author[[7]]
#> [[1]]$author[[7]][[1]]
#> [1] "M. C."
#> 
#> [[1]]$author[[7]][[2]]
#> [1] "Garcia-Pelayo"
#> 
#> 
#> [[1]]$author[[8]]
#> [[1]]$author[[8]][[1]]
#> [1] "P. J."
#> 
#> [[1]]$author[[8]][[2]]
#> [1] "Hogarth"
#> 
#> 
#> [[1]]$author[[9]]
#> [[1]]$author[[9]][[1]]
#> [1] "H."
#> 
#> [[1]]$author[[9]][[2]]
#> [1] "McShane"
#> 
#> 
#> 
#> [[1]]$PMID
#> [1] "30446726"
#> 
#> [[1]]$PMCID
#> [1] "7051908"
#> 
#> [[1]]$source
#> [[1]]$source[[1]]
#> [1] "PubMed"
#> 
#> [[1]]$source[[2]]
#> [1] "Zotero"
#> 
#> 
#> 

if (FALSE) {
doi1 <- "10.1108/jd-12-2013-0166"
doi2 <- "10.1371/journal.pone.0058568"
pmid1 <- 30446726
pmcid1 <- "PMC4679344"
isbn1 <- 1439895619

# different formats
cit_oid(doi1)
cit_oid(pmid1, format = "mediawiki")
cit_oid(pmid1, format = "zotero")
cit_oid(pmid1, format = "mediawiki-basefields")
cat(cit_oid(pmid1, format = "bibtex")[[1]])

# PMID example
cit_oid(pmid1, verbose = TRUE)

# ISBN example
cit_oid(isbn1, verbose = TRUE)

# PMCID example
cit_oid(pmcid1)

# set the accept language
x <- cit_oid(pmid1, accept_language = "fr-FR", verbose = TRUE)
x <- cit_oid(doi2, accept_language = "de-DE", verbose = TRUE)

# just get raw text/json
cit_oid_(pmcid1)

# many ids at once
cit_oid(id = c(pmid1, pmcid1, isbn1))
cit_oid_(id = c(pmid1, pmcid1, isbn1))
cit_oid_(id = c(pmid1, pmcid1, isbn1), format = "bibtex")
}